+++ src/components/button/button-static/button-static.spec.ts
@@ -86,15 +86,4 @@ describe('sbb-button-static', () => {
     expect(dataSlots).to.contain('icon');
     expect(dataSlots).not.to.contain('unnamed');
   });
-
-  it('should render form field button variant when inside of a form field', async () => {
-    const root = await fixture(
-      html` <sbb-form-field>
-        <input />
-        <sbb-button-static slot="suffix" icon-name="cross-small"></sbb-button-static>
-      </sbb-form-field>`,
-    );
-    const button = root.querySelector('sbb-button-static');
-    expect(button).to.have.attribute('data-icon-small');
-  });
 });
