+++ src/components/button/mini-button-link/mini-button-link.ts
@@ -1,27 +1,31 @@
-import type { CSSResultGroup } from 'lit';
+import type { CSSResultGroup, TemplateResult } from 'lit';
 import { customElement } from 'lit/decorators.js';
 
 import { SbbDisabledTabIndexActionMixin, SbbLinkBaseElement } from '../../core/common-behaviors';
 import { SbbButtonCommonElementMixin } from '../common/button-common';
 import commonStyle from '../common/button-common.scss?lit&inline';
-import style from '../common/primary-button.scss?lit&inline';
+import style from '../common/mini-button.scss?lit&inline';
 
 /**
- * It displays a button enhanced with the SBB Design in the 'transparent' variant, which will behave as a link.
+ * It displays an icon-only button enhanced with the SBB Design, which will behave as a link;
+ * it's meant to be used mainly within the sbb-form-field in prefix/suffix slot.
  *
- * @slot - Use the unnamed slot to add content to the button-link.
  * @slot icon - Slot used to display the icon, if one is set
  */
-@customElement('sbb-transparent-button-link')
-export class SbbTransparentButtonLinkElement extends SbbButtonCommonElementMixin(
+@customElement('sbb-mini-button-link')
+export class SbbMiniButtonLinkElement extends SbbButtonCommonElementMixin(
   SbbDisabledTabIndexActionMixin(SbbLinkBaseElement),
 ) {
   public static override styles: CSSResultGroup = [commonStyle, style];
+
+  protected override renderTemplate(): TemplateResult {
+    return this.renderIcon();
+  }
 }
 
 declare global {
   interface HTMLElementTagNameMap {
     // eslint-disable-next-line @typescript-eslint/naming-convention
-    'sbb-transparent-button-link': SbbTransparentButtonLinkElement;
+    'sbb-mini-button-link': SbbMiniButtonLinkElement;
   }
 }
