+++ src/components/button/mini-button/index.ts
@@ -1,0 +1,1 @@
+export * from './mini-button';
