+++ src/components/form-field/form-field/form-field.ts
@@ -464,7 +464,7 @@ export class SbbFormFieldElement extends SbbNegativeMixin(LitElement) {
 
   private _syncNegative(): void {
     this.querySelectorAll?.(
-      'sbb-form-error,sbb-button,sbb-popover-trigger,sbb-form-field-clear,sbb-datepicker-next-day,sbb-datepicker-previous-day,sbb-datepicker-toggle,sbb-select,sbb-autocomplete',
+      'sbb-form-error,sbb-mini-button,sbb-mini-button-link,sbb-popover-trigger,sbb-form-field-clear,sbb-datepicker-next-day,sbb-datepicker-previous-day,sbb-datepicker-toggle,sbb-select,sbb-autocomplete',
     ).forEach((element) =>
       this.negative ? element.setAttribute('negative', '') : element.removeAttribute('negative'),
     );
